various tweaks
- move index.js to lib folder
- observe config instead of reading config each time
