+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.5.1",
   "private": true,
   "description": "A plugin for Atom Linter providing an interface to textlint.",
+  "main": "lib/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/1000ch/linter-textlint.git"
@@ -29,14 +30,14 @@
   },
   "homepage": "https://github.com/1000ch/linter-textlint#readme",
   "dependencies": {
-    "atom-linter": "^4.3.0",
+    "atom-linter": "^4.3.2",
     "atom-package-deps": "^3.0.7",
     "textlint": "^5.2.0"
   },
   "devDependencies": {
     "babel-eslint": "^4.1.6",
     "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^3.0.0",
+    "eslint-config-airbnb": "^3.1.0",
     "textlint-rule-preset-japanese": "^1.0.2",
     "textlint-rule-preset-jtf-style": "^2.0.0"
   },
