+++ spec/linter-textlint-spec.js
@@ -3,7 +3,7 @@
 import * as path from 'path';
 
 describe('The textlint provider for Linter', () => {
-  const lint = require(path.join('..', 'index.js')).provideLinter().lint;
+  const lint = require(path.join('..', 'lib', 'index.js')).provideLinter().lint;
 
   beforeEach(() => {
     atom.workspace.destroyActivePaneItem();
@@ -40,13 +40,9 @@ describe('The textlint provider for Linter', () => {
     it('verifies the first message', () => {
       waitsForPromise(() => {
         return lint(editor).then(messages => {
-          expect(messages[0].type).toBeDefined();
           expect(messages[0].type).toEqual('Error');
-          expect(messages[0].text).toBeDefined();
           expect(messages[0].text).toEqual('一文に二回以上利用されている助詞 "て" がみつかりました。');
-          expect(messages[0].filePath).toBeDefined();
           expect(messages[0].filePath).toMatch(/.+bad\.md$/);
-          expect(messages[0].range).toBeDefined();
           expect(messages[0].range).toEqual({
             start: { row: 0, column: 10 },
             end: { row: 0, column: 10 }
